changed

+++ src/main/java/com/example/project11/Controllers/Charts/BarChartController.java
@@ -27,7 +27,7 @@ public class BarChartController {
             "#89CFF0"
     };
 
-    public void setChartData(Map<String, Double> gradeDistribution) {
+    public void setChartData(Map<String, Integer> gradeDistribution) {
 
         barChart.getData().clear();
 
@@ -36,7 +36,7 @@ public class BarChartController {
 
         // Prepare color list and populate the series with data
         List<String> colorList = new ArrayList<>();
-        for (Map.Entry<String, Double> entry : gradeDistribution.entrySet()) {
+        for (Map.Entry<String, Integer> entry : gradeDistribution.entrySet()) {
             // Add each entry as a new Data point in the series
             series.getData().add(new XYChart.Data<>(entry.getKey(), entry.getValue()));
             colorList.add(colors[colorList.size() % colors.length]); // Cycles through colors
